move sensor related code into AbstractCamera
