+++ release/files_tici
@@ -8,6 +8,7 @@ selfdrive/assets/navigation/*
 selfdrive/assets/training_wide/*
 
 system/camerad/cameras/camera.h
+system/camerad/cameras/camera.cc
 system/camerad/cameras/camera_ar0231.cc
 system/camerad/cameras/camera_ox03c10.cc
 system/camerad/cameras/camera_qcom2.cc
