+++ system/camerad/cameras/camera.h
@@ -8,12 +8,14 @@
 #include <vector>
 
 #include "cereal/messaging/messaging.h"
+#include "common/util.h"
 #include "media/cam_defs.h"
 #include "media/cam_isp.h"
 #include "media/cam_isp_ife.h"
 #include "media/cam_sensor.h"
 #include "media/cam_sensor_cmn_header.h"
 #include "media/cam_sync.h"
+#include "system/camerad/cameras/camera_util.h"
 #include "system/camerad/cameras/sensor2_i2c.h"
 
 #define CAMERA_ID_IMX298 0
@@ -40,16 +42,30 @@ typedef struct CameraInfo {
   int stats_offset = -1;
 } CameraInfo;
 
+class MultiCameraState;
+
 class AbstractCamera {
 public:
   AbstractCamera() = default;
   virtual ~AbstractCamera(){};
-  virtual int getSlaveAddress(int port) const = 0;
+  int open(MultiCameraState *multi_cam_state, int camera_num, bool enabled);
+  void close();
+  void sensors_start();
+  void sensors_i2c(const std::vector<i2c_random_wr_payload> &dat, int op_code);
+  void sensors_poke(int request_id);
+  void config_isp(int io_mem_handle, int fence, int request_id, int buf0_mem_handle, int buf0_offset);
   virtual std::vector<i2c_random_wr_payload> getExposureVector(int new_g, bool dc_gain_enabled, int exposure_time, int dc_gain_weight) const = 0;
   virtual void processRegisters(void *addr, cereal::FrameData::Builder &framed) const {}
 
+  int camera_id;
+  int camera_num;
+  bool enabled;
+
+  int buf0_handle;
+  int32_t link_handle;
+  int32_t session_handle;
   CameraInfo ci;
-  int id;
+
   float dc_gain_factor;
   int dc_gain_min_weight;
   int dc_gain_max_weight;
@@ -64,6 +80,18 @@ public:
   float min_ev, max_ev;
   int registers_offset = 0;
   float sensor_analog_gains[35];
+
+protected:
+  int sensors_init();
+  virtual int getSlaveAddress(int port) const = 0;
+
+  MultiCameraState *multi_cam_state;
+  unique_fd sensor_fd;
+  unique_fd csiphy_fd;
+  int32_t sensor_dev_handle;
+  int32_t isp_dev_handle;
+  int32_t csiphy_dev_handle;
+  MemoryManager mm;
 
   camera_sensor_i2c_type i2c_type;
   std::vector<struct i2c_random_wr_payload> start_reg_array;
