+++ system/camerad/cameras/camera_common.cc
@@ -37,7 +37,7 @@ public:
              "-DIS_OX=%d -DCAM_NUM=%d%s",
              ci->frame_width, ci->frame_height, ci->frame_stride, ci->frame_offset,
              b->rgb_width, b->rgb_height, b->rgb_stride, buf_width, uv_offset,
-             s->camera->id==CAMERA_ID_OX03C10 ? 1 : 0, s->camera_num, s->camera_num==1 ? " -DVIGNETTING" : "");
+             s->camera->camera_id==CAMERA_ID_OX03C10 ? 1 : 0, s->camera->camera_num, s->camera->camera_num==1 ? " -DVIGNETTING" : "");
     const char *cl_file = "cameras/real_debayer.cl";
     cl_program prg_debayer = cl_program_from_file(context, device_id, cl_file, args);
     krnl_ = CL_CHECK_ERR(clCreateKernel(prg_debayer, "debayer10", &err));
@@ -166,9 +166,9 @@ void fill_frame_data(cereal::FrameData::Builder &framed, const FrameMetadata &fr
   const float perc = util::map_val(ev, c->camera->min_ev, c->camera->max_ev, 0.0f, 100.0f);
   framed.setExposureValPercent(perc);
 
-  if (c->camera->id == CAMERA_ID_AR0231) {
+  if (c->camera->camera_id == CAMERA_ID_AR0231) {
     framed.setSensor(cereal::FrameData::ImageSensor::AR0321);
-  } else if (c->camera->id == CAMERA_ID_OX03C10) {
+  } else if (c->camera->camera_id == CAMERA_ID_OX03C10) {
     framed.setSensor(cereal::FrameData::ImageSensor::OX03C10);
   }
 }
